Modify codestyle and remove redundant dependency
