+++ pom.xml
@@ -7,7 +7,7 @@
     <groupId>com.microsoft.azure</groupId>
     <artifactId>spring-jms-service-bus</artifactId>
     <packaging>pom</packaging>
-    <version>1.0-SNAPSHOT</version>
+    <version>1.0.0-SNAPSHOT</version>
 
     <modules>
         <module>spring-jms-servicebus-sample</module>
