+++ spring-jms-servicebus-sample/azure-servicebus-queue-spring-jms-sample/pom.xml
@@ -5,8 +5,8 @@
 
     <parent>
         <groupId>com.microsoft.azure</groupId>
-        <artifactId>azure-servicebus-spring-boot-jms-sample</artifactId>
-        <version>1.0-SNAPSHOT</version>
+        <artifactId>spring-jms-service-bus-sample</artifactId>
+        <version>1.0.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
@@ -38,14 +38,6 @@
             <version>5.1.8.RELEASE</version>
         </dependency>
         <!--End Spring JMS-->
-
-        <!--Azure Service Bus-->
-        <dependency>
-            <groupId>com.microsoft.azure</groupId>
-            <artifactId>azure-servicebus</artifactId>
-            <version>1.2.12</version>
-        </dependency>
-        <!--End Azure Service Bus-->
 
         <!--Qpid-->
         <dependency>
