+++ spring-jms-servicebus-sample/azure-servicebus-queue-spring-jms-sample/src/main/java/sample/ServiceBusJMSQueueApplication.java
@@ -9,10 +9,47 @@ package sample;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.jms.annotation.EnableJms;
+import org.apache.qpid.jms.JmsConnectionFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.jms.config.DefaultJmsListenerContainerFactory;
+import org.springframework.jms.config.JmsListenerContainerFactory;
+import org.springframework.jms.connection.CachingConnectionFactory;
+import org.springframework.beans.factory.annotation.Value;
+import javax.jms.ConnectionFactory;
 
 @SpringBootApplication
 @EnableJms
 public class ServiceBusJMSQueueApplication {
+
+    @Value("${spring.jms.servicebus.connection-string}")
+    private String connectionString;
+
+    @Value("${spring.jms.servicebus.idle-timeout}")
+    private int idleTimeout;
+
+    private static final String AMQP_URI_FORMAT = "amqps://%s?amqp.idleTimeout=%d";
+
+    @Bean
+    public ConnectionFactory myConnectionFactory() {
+        ServiceBusKey serviceBusKey = ConnectionStringResolver.getServiceBusKey(connectionString);
+        String host = serviceBusKey.getHost();
+        String sasKeyName = serviceBusKey.getSharedAccessKeyName();
+        String sasKey = serviceBusKey.getSharedAccessKey();
+
+        String remoteUri = String.format(AMQP_URI_FORMAT, host, idleTimeout);
+        JmsConnectionFactory jmsConnectionFactory = new JmsConnectionFactory();
+        jmsConnectionFactory.setRemoteURI(remoteUri);
+        jmsConnectionFactory.setUsername(sasKeyName);
+        jmsConnectionFactory.setPassword(sasKey);
+        return new CachingConnectionFactory(jmsConnectionFactory);
+    }
+
+    @Bean
+    public JmsListenerContainerFactory<?> myQueueFactory(ConnectionFactory connectionFactory) {
+        DefaultJmsListenerContainerFactory queueFactory = new DefaultJmsListenerContainerFactory();
+        queueFactory.setConnectionFactory(connectionFactory);
+        return queueFactory;
+    }
 
     public static void main(String[] args) {
         SpringApplication.run(ServiceBusJMSQueueApplication.class, args);
