+++ spring-jms-servicebus-sample/azure-servicebus-queue-spring-jms-sample/src/main/java/sample/User.java
@@ -11,6 +11,7 @@ import java.io.Serializable;
 public class User implements Serializable {
 
     private static final long serialVersionUID = -295422703255886286L;
+
     private String name;
 
     public User(String name) {
