+++ spring-jms-servicebus-sample/azure-servicebus-topic-spring-jms-sample/src/main/java/sample/ServiceBusKey.java
@@ -7,8 +7,11 @@
 package sample;
 
 public class ServiceBusKey {
+
     private final String host;
+
     private final String sharedAccessKeyName;
+
     private final String sharedAccessKey;
 
     ServiceBusKey(String host, String sharedAccessKeyName, String sharedAccessKey) {
