+++ spring-jms-servicebus-sample/azure-servicebus-topic-spring-jms-sample/src/main/java/sample/TopicReceiveController.java
@@ -9,7 +9,6 @@ package sample;
 import org.apache.qpid.jms.JmsConnectionFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.jms.annotation.JmsListener;
 import org.springframework.jms.config.DefaultJmsListenerContainerFactory;
@@ -22,46 +21,11 @@ import javax.jms.ConnectionFactory;
 @Component
 public class TopicReceiveController {
 
-    @Value("${spring.jms.servicebus.connection-string}")
-    private String connectionString;
+    private static final String TOPIC_NAME = "testtopic";
 
-    @Value("${spring.jms.servicebus.topic-client-id}")
-    private String clientId = "";
-
-    @Value("${spring.jms.servicebus.idle-timeout}")
-    private int idleTimeout;
-
-    private static final String AMQP_URI_FORMAT = "amqps://%s?amqp.idleTimeout=%d";
-
-    private static final String TOPIC_NAME = "tpc001";
-
-    private static final String SUBSCRIPTION_NAME = "sub001";
+    private static final String SUBSCRIPTION_NAME = "testsubscription";
 
     private final Logger logger = LoggerFactory.getLogger(TopicReceiveController.class);
-
-    @Bean
-    public ConnectionFactory myConnectionFactory() {
-        ServiceBusKey serviceBusKey = ConnectionStringResolver.getServiceBusKey(connectionString);
-        String host = serviceBusKey.getHost();
-        String sasKeyName = serviceBusKey.getSharedAccessKeyName();
-        String sasKey = serviceBusKey.getSharedAccessKey();
-
-        String remoteUri = String.format(AMQP_URI_FORMAT, host, idleTimeout);
-        JmsConnectionFactory jmsConnectionFactory = new JmsConnectionFactory();
-        jmsConnectionFactory.setRemoteURI(remoteUri);
-        jmsConnectionFactory.setClientID(clientId);
-        jmsConnectionFactory.setUsername(sasKeyName);
-        jmsConnectionFactory.setPassword(sasKey);
-        return new CachingConnectionFactory(jmsConnectionFactory);
-    }
-
-    @Bean
-    public JmsListenerContainerFactory<?> myTopicFactory(ConnectionFactory connectionFactory) {
-        DefaultJmsListenerContainerFactory topicFactory = new DefaultJmsListenerContainerFactory();
-        topicFactory.setConnectionFactory(connectionFactory);
-        topicFactory.setSubscriptionDurable(Boolean.TRUE);
-        return topicFactory;
-    }
 
     @JmsListener(destination = TOPIC_NAME, containerFactory = "myTopicFactory", subscription = SUBSCRIPTION_NAME)
     public void receiveMessage(User user) {
