+++ spring-jms-servicebus-sample/pom.xml
@@ -4,9 +4,9 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.microsoft.azure</groupId>
-    <artifactId>azure-servicebus-spring-boot-jms-sample</artifactId>
+    <artifactId>spring-jms-service-bus-sample</artifactId>
     <packaging>pom</packaging>
-    <version>1.0-SNAPSHOT</version>
+    <version>1.0.0-SNAPSHOT</version>
 
     <modules>
         <module>azure-servicebus-queue-spring-jms-sample</module>
